Move FITS tests to separate test file
- Extract inline tests from fits.rs to tests/test_fits.rs
- Add base64 as dev-dependency for test file
- Keep connection.rs tests inline (pub(crate) module)
