+++ services/phd2-guider/Cargo.toml
@@ -20,3 +20,4 @@ fitsio = { workspace = true }
 
 [dev-dependencies]
 tokio-test = { workspace = true }
+base64 = { workspace = true }
